chore: merge new marketplace servce into one
