+++ apps/ui/src/hooks/marketplace/useGetCollection.ts
@@ -1,8 +1,8 @@
 import { MarketplaceQueryKeys } from '@/utils/constants';
 import { useChainId } from '../useChainId';
-import { newMarketplaceService } from '@/services/new-marketplace';
 import { useQuery } from '@tanstack/react-query';
 import { Networks } from '@/utils/resolverNetwork';
+import { marketplaceService } from '@/services/marketplace';
 
 type useGetCollectionProps = {
   collectionId: string;
@@ -14,7 +14,7 @@ export const useGetCollection = ({ collectionId }: useGetCollectionProps) => {
   const { data: collection, ...rest } = useQuery({
     queryKey: [MarketplaceQueryKeys.COLLECTION, chainId, collectionId],
     queryFn: async () => {
-      const { data } = await newMarketplaceService.getCollection({
+      const { data } = await marketplaceService.getCollection({
         collectionId,
         chainId: chainId ?? Networks.MAINNET,
       });
