+++ apps/ui/src/hooks/marketplace/useGetCollectionOrders.ts
@@ -2,7 +2,7 @@ import { MarketplaceQueryKeys } from '@/utils/constants';
 import type { PaginationResult } from '@/utils/pagination';
 import { useInfiniteQuery } from '@tanstack/react-query';
 import { useChainId } from '../useChainId';
-import { newMarketplaceService } from '@/services/new-marketplace';
+import { marketplaceService } from '@/services/marketplace';
 import { Networks } from '@/utils/resolverNetwork';
 import type { Order } from '@/types/marketplace';
 
@@ -43,7 +43,7 @@ export const useGetCollectionOrders = ({
       return undefined;
     },
     queryFn: async ({ pageParam = 0 }) => {
-      const { data } = await newMarketplaceService.getCollectionOrders({
+      const { data } = await marketplaceService.getCollectionOrders({
         page: pageParam as number,
         limit,
         search,
