+++ apps/ui/src/hooks/marketplace/useGetOrder.ts
@@ -1,7 +1,7 @@
 import { MarketplaceQueryKeys } from '@/utils/constants';
 import { useQuery } from '@tanstack/react-query';
 import { useChainId } from '../useChainId';
-import { newMarketplaceService } from '@/services/new-marketplace';
+import { marketplaceService } from '@/services/marketplace';
 import { Networks } from '@/utils/resolverNetwork';
 
 type useGetOrderProps = { id: string };
@@ -12,7 +12,7 @@ export const useGetOrder = ({ id }: useGetOrderProps) => {
   const { data: order, ...rest } = useQuery({
     queryKey: [MarketplaceQueryKeys.ORDER, chainId, id],
     queryFn: async () => {
-      const { data } = await newMarketplaceService.getOrder({
+      const { data } = await marketplaceService.getOrder({
         orderId: id,
         chainId: chainId ?? Networks.MAINNET,
       });
