+++ apps/ui/src/hooks/marketplace/useListCollections.ts
@@ -2,9 +2,9 @@ import { MarketplaceQueryKeys } from '@/utils/constants';
 import type { PaginationResult } from '@/utils/pagination';
 import { useInfiniteQuery } from '@tanstack/react-query';
 import { useChainId } from '../useChainId';
-import { newMarketplaceService } from '@/services/new-marketplace';
 import { Networks } from '@/utils/resolverNetwork';
 import type { Collection } from '@/types/marketplace';
+import { marketplaceService } from '@/services/marketplace';
 
 type useGetCollectionsProps = {
   page?: number;
@@ -41,7 +41,7 @@ export const useGetCollections = ({
       return undefined;
     },
     queryFn: async ({ pageParam = 1 }) => {
-      const { data } = await newMarketplaceService.listCollections({
+      const { data } = await marketplaceService.listCollections({
         page: pageParam as number,
         limit,
         search,
