+++ apps/ui/src/services/marketplace.ts
@@ -1,30 +1,19 @@
-import type { Asset } from '@/types/marketplace';
+import type {
+  Asset,
+  Collection,
+  MarketplacePaginatedResponse,
+  MarketplacePaginatedResponseUserOrders,
+  Order,
+  OrderWithMedatada,
+} from '@/types/marketplace';
 import { constructUrl } from '@/utils/constructUrl';
-import type { OrderResponse } from '@/utils/getOrderMetadata';
 import { Networks, resolveNetwork } from '@/utils/resolverNetwork';
 
 const BASE_API_URL = import.meta.env.VITE_API_URL;
 
+const BASE_MARKETPLACE_URL = import.meta.env.VITE_MARKETPLACE_URL;
+
 export class marketplaceService {
-  static async getOrdersByAccount({
-    account,
-    page,
-    chainId = Networks.MAINNET,
-  }: { account: string; page: number | string; chainId?: number }): Promise<{
-    orders: OrderResponse[];
-    total: number;
-  }> {
-    const limit = 12;
-    const network = resolveNetwork(chainId);
-    const response = await fetch(
-      `${BASE_API_URL}/${network}/marketplace/${account}/orders?page=${page}&limit=${limit}`
-    );
-
-    const data = await response.json();
-
-    return data;
-  }
-
   static async getAssets({
     chainId = Networks.MAINNET,
   }: { chainId?: number }): Promise<Omit<Asset, 'metadata'>[]> {
@@ -38,25 +27,107 @@ export class marketplaceService {
     return data;
   }
 
-  static async getOrders({
+  static async getCollection({
+    collectionId,
+    chainId = Networks.MAINNET,
+  }: {
+    collectionId: string;
+    chainId?: number;
+  }): Promise<{ data: Collection }> {
+    const network = resolveNetwork(chainId);
+
+    const url = constructUrl(
+      `${BASE_MARKETPLACE_URL}/${network}/collections/${collectionId}`,
+      {}
+    );
+
+    const response = await fetch(url);
+
+    const data = await response.json();
+
+    return data;
+  }
+  static async getOrder({
+    orderId,
+    chainId = Networks.MAINNET,
+  }: {
+    orderId: string;
+    chainId?: number;
+  }): Promise<{ data: OrderWithMedatada }> {
+    const network = resolveNetwork(chainId);
+
+    const url = constructUrl(
+      `${BASE_MARKETPLACE_URL}/${network}/orders/${orderId}`,
+      {}
+    );
+
+    const response = await fetch(url);
+
+    const data = await response.json();
+
+    return data;
+  }
+
+  static async getCollectionOrders({
+    collectionId,
     page,
     limit,
     search,
     chainId = Networks.MAINNET,
+    sortValue,
+    sortDirection,
+  }: {
+    collectionId: string;
+    page: number | string;
+    limit: number;
+    search?: string;
+    chainId?: number;
+    sortValue: string;
+    sortDirection: 'asc' | 'desc';
+  }): Promise<MarketplacePaginatedResponse<Order>> {
+    const network = resolveNetwork(chainId);
+
+    const url = constructUrl(
+      `${BASE_MARKETPLACE_URL}/${network}/collections/${collectionId}/orders`,
+      {
+        page,
+        limit,
+        assetId: search,
+        orderBy: sortValue,
+        orderDirection: sortDirection,
+      }
+    );
+
+    const response = await fetch(url);
+
+    const data = await response.json();
+
+    return data;
+  }
+
+  static async listCollections({
+    page,
+    limit,
+    search,
+    chainId = Networks.MAINNET,
+    sortValue,
+    sortDirection,
   }: {
     page: number | string;
     limit: number;
     search?: string;
     chainId?: number;
-  }): Promise<{
-    orders: OrderResponse[];
-    total: number;
-  }> {
+    sortValue: string;
+    sortDirection: 'asc' | 'desc';
+  }): Promise<MarketplacePaginatedResponse<Collection>> {
     const network = resolveNetwork(chainId);
-    const url = constructUrl(`${BASE_API_URL}/${network}/marketplace/orders`, {
+
+    const url = constructUrl(`${BASE_MARKETPLACE_URL}/${network}/collections`, {
       page,
       limit,
-      search,
+      name: search,
+      orderBy: sortValue,
+      orderDirection: sortDirection,
     });
 
     const response = await fetch(url);
@@ -66,49 +137,34 @@ export class marketplaceService {
     return data;
   }
 
-  static async getOrderById({
-    id,
+  static async listUserOrders({
+    page,
     chainId = Networks.MAINNET,
+    sellerAddress,
+    limit,
   }: {
-    id: string;
-    chainId?: number;
-  }): Promise<OrderResponse | null> {
-    try {
-      const network = resolveNetwork(chainId);
-      const response = await fetch(
-        `${BASE_API_URL}/${network}/marketplace/orders/${id}`
-      );
+    page: number | string;
+    chainId: number;
+    limit?: number;
+    sellerAddress: string;
+  }): Promise<MarketplacePaginatedResponseUserOrders<Order>> {
+    const network = resolveNetwork(chainId);
 
-      const data = await response.json();
-
-      if (response.status !== 200) {
-        return null;
+    const url = constructUrl(
+      `${BASE_MARKETPLACE_URL}/${network}/user/orders/${sellerAddress}`,
+      {
+        page,
+        limit: limit ?? 10,
+        orderBy: 'createdAt',
+        orderDirection: 'desc',
+        sellerAddress,
       }
+    );
 
-      return data;
-    } catch {
-      return null;
-    }
-  }
+    const response = await fetch(url);
 
-  static async getAssetById({
-    id,
-    chainId = Networks.MAINNET,
-  }: {
-    id: string;
-    chainId?: number;
-  }): Promise<Asset | null> {
-    try {
-      const network = resolveNetwork(chainId);
-      const response = await fetch(
-        `${BASE_API_URL}/${network}/marketplace/assets/${id}`
-      );
+    const data = await response.json();
 
-      const data = await response.json();
-
-      return data;
-    } catch {
-      return null;
-    }
+    return data;
   }
 }
